Extract market status into separate class
